+++ src/scrapers/document/DocumentScraper.test.ts
@@ -1,20 +1,22 @@
 import { describe, expect, it } from '@jest/globals';
 import { join } from 'path';
+import { $provideExecutablesForNode } from '../../execution/utils/$provideExecutablesForNode';
 import { $provideLlmToolsForTestingAndScriptsAndPlayground } from '../../llm-providers/_common/register/$provideLlmToolsForTestingAndScriptsAndPlayground';
 import { $provideFilesystemForNode } from '../_common/register/$provideFilesystemForNode';
 import { makeKnowledgeSourceHandler } from '../_common/utils/makeKnowledgeSourceHandler';
 import { DocumentScraper } from './DocumentScraper';
 
-describe('how creating knowledge from docx works', () => {
+describe('how creating knowledge from docx works', async () => {
     const rootDirname = join(__dirname, 'samples');
+
     const documentScraper = new DocumentScraper(
-        { fs: $provideFilesystemForNode(), llm: $provideLlmToolsForTestingAndScriptsAndPlayground() },
+        {
+            fs: $provideFilesystemForNode(),
+            llm: $provideLlmToolsForTestingAndScriptsAndPlayground(),
+            executables: await $provideExecutablesForNode(),
+        },
         {
             rootDirname,
-            externalProgramsPaths: {
-                // TODO: !!!!!! use `locate-app` library here
-                pandocPath: 'C:/Users/me/AppData/Local/Pandoc/pandoc.exe',
-            },
         },
     );