+++ CHANGELOG.md
@@ -540,6 +540,9 @@ Knowledge scrapers [🐝]
 -   Change `FileStorage` -> `FileCacheStorage`
 -   Changed behavior of `titleToName` when passing URLs or file paths
 -   Fix normalize functions when normalizing string containing slash char "/", "\"
+-   Pass `fs` through `ExecutionTools`
+-   Pass `executables` through `ExecutionTools`
+-   Pass `scrapers` through `ExecutionTools`
 -   Add utilities `$provideExecutionToolsForBrowser` and `$provideExecutionToolsForNode` and use them in samples
 -   Add utilities `$provideScrapersForBrowser` and `$provideScrapersForNode`
 -   Rename `createLlmToolsFromConfigurationFromEnv` -> `$provideLlmToolsConfigurationFromEnv` and `createLlmToolsFromEnv` -> `$provideLlmToolsFromEnv`